requirements-dev.txt
-r requirements.txt
black
pytest
checkov
awscurl
python-dotenv
